fix bug pt9, bug route from middleware cashier and owner, and add some validate

+++ resources/views/kategori/data.blade.php
@@ -14,16 +14,29 @@
                     <td class="text-center">{{ $no++ }}</td>
                     <td>{{ $item->nama }}</td>
                     <td>
-                        <button class="btn btn-xs btn-warning" type="button" style="color: green;" title="Edit" data-mode="edit" data-toggle="modal" data-target="#formModalKategori" data-id_kategori="{{ $item->id }}" data-nama_kategori="{{ $item->nama }}" data-route="{{ route('admin.kategori.update', $item->id) }}">
-                            <i class="fas fa-edit"></i>
-                        </button>
-                        <form action="{{ route('admin.kategori.destroy', $item->id) }}" style="display: inline;" method="post">
-                            @csrf
-                            @method('DELETE')
-                            <button onclick="deleteData('{{ route('admin.kategori.destroy', $item->id) }}')" class="btn btn-xs btn-danger delete-data" type="button" title="Delete">
-                                <i class="fas fa-trash"></i>
+                        @if (auth()->user()->hak_akses == 'admin')
+                            <button class="btn btn-xs btn-warning" type="button" style="color: green;" title="Edit" data-mode="edit" data-toggle="modal" data-target="#formModalKategori" data-id_kategori="{{ $item->id }}" data-nama_kategori="{{ $item->nama }}" data-route="{{ route('admin.kategori.update', $item->id) }}">
+                                <i class="fas fa-edit"></i>
                             </button>
-                        </form>
+                            <form action="{{ route('admin.kategori.destroy', $item->id) }}" style="display: inline;" method="post">
+                                @csrf
+                                @method('DELETE')
+                                <button onclick="deleteData('{{ route('admin.kategori.destroy', $item->id) }}')" class="btn btn-xs btn-danger delete-data" type="button" title="Delete">
+                                    <i class="fas fa-trash"></i>
+                                </button>
+                            </form>
+                        @elseif (auth()->user()->hak_akses == 'owner')
+                            <button class="btn btn-xs btn-warning" type="button" style="color: green;" title="Edit" data-mode="edit" data-toggle="modal" data-target="#formModalKategori" data-id_kategori="{{ $item->id }}" data-nama_kategori="{{ $item->nama }}" data-route="{{ route('owner.kategori.update', $item->id) }}">
+                                <i class="fas fa-edit"></i>
+                            </button>
+                            <form action="{{ route('owner.kategori.destroy', $item->id) }}" style="display: inline;" method="post">
+                                @csrf
+                                @method('DELETE')
+                                <button onclick="deleteData('{{ route('owner.kategori.destroy', $item->id) }}')" class="btn btn-xs btn-danger delete-data" type="button" title="Delete">
+                                    <i class="fas fa-trash"></i>
+                                </button>
+                            </form>
+                        @endif
                     </td>
                 </tr>
             @endforeach
